separate the add_constraint method from the constraint class
