+++ db/tests/constraints/operations/test_drop.py
@@ -1,6 +1,6 @@
 from sqlalchemy import String, Integer, Column, Table, MetaData
-
-from db.constraints.operations.create import UniqueConstraint
+from db.constraints.base import UniqueConstraint
+from db.constraints.operations.create import add_constraint
 from db.constraints.operations.drop import drop_constraint
 from db.tables.operations.select import get_oid_from_table, reflect_table_from_oid
 from db.tests.constraints import utils as test_utils
@@ -21,7 +21,7 @@ def test_drop_unique_constraint(engine_with_schema):
 
     table_oid = get_oid_from_table(table_name, schema, engine)
     uq_constraint = UniqueConstraint(None, table_oid, [unique_column_name])
-    uq_constraint.add_constraint(engine)
+    add_constraint(uq_constraint, engine)
     altered_table = reflect_table_from_oid(table_oid, engine, metadata=get_empty_metadata())
     test_utils.assert_primary_key_and_unique_present(altered_table)
     unique_constraint = test_utils.get_first_unique_constraint(altered_table)
